backend/requirements.txt
Django==1.8
djangorestframework==3.4.0
MySQL-python==1.2.5
pkg-resources==0.0.0
